chore: fix useEffect is not defined, turn off autoimports in wxt (#123)
The default wxt config uses unimport, which automatically sets up certain imports in the project. However, on the first load of the extension after reloading the it on the extensions page, it errors out because the imports can't be found properly. Adding the imports fixes this.

Also, biome needs the import to be able to detect issues like "hook does not list all dependencies", etc.